requirements.txt
google-genai
matplotlib
pandas
python-dotenv
streamlit
